# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: beelms API
  version: 0.1.0
  license:
    name: UNLICENSED
    url: https://example.com/unlicense
  description: |
    OpenAPI спецификация за основните MVP услуги на beelms.
servers:
  - url: https://api.beelms.com/api
    description: Production (примерен домейн, с /api префикс)
  - url: http://127.0.0.1:3000/api
    description: Local development (с /api префикс)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PublicSettings:
      type: object
      properties:
        branding:
          type: object
          properties:
            appName:
              type: string
            logoUrl:
              type: string
              nullable: true
            primaryColor:
              type: string
              nullable: true
          required: [appName]
        features:
          type: object
          properties:
            wikiPublic:
              type: boolean
            courses:
              type: boolean
            auth:
              type: boolean
            paidCourses:
              type: boolean
            gdprLegal:
              type: boolean
            infraRedis:
              type: boolean
            infraRabbitmq:
              type: boolean
            infraMonitoring:
              type: boolean
            infraErrorTracking:
              type: boolean
          required:
            [wikiPublic, courses, auth, paidCourses, gdprLegal, infraRedis, infraRabbitmq, infraMonitoring, infraErrorTracking]
        languages:
          type: object
          properties:
            supported:
              type: array
              items:
                type: string
            default:
              type: string
          required: [supported, default]
      required: [branding, features, languages]
...
    AdminUpdateInstanceSettingsInput:
      type: object
      properties:
        branding:
          type: object
          properties:
            appName:
              type: string
            logoUrl:
              type: string
              nullable: true
            primaryColor:
              type: string
              nullable: true
        features:
          type: object
          properties:
            wikiPublic:
              type: boolean
            courses:
              type: boolean
            auth:
              type: boolean
            paidCourses:
              type: boolean
            gdprLegal:
              type: boolean
            infraRedis:
              type: boolean
            infraRabbitmq:
              type: boolean
            infraMonitoring:
              type: boolean
            infraErrorTracking:
              type: boolean
        languages:
          type: object
          properties:
            supported:
              type: array
              items:
                type: string
            default:
              type: string
...
security:
  - bearerAuth: []
paths:
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [Admin]
      summary: Current instance settings (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current instance settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettings'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateAdminSettings
      tags: [Admin]
      summary: Update instance settings (admin)
      description: Partial update (PATCH) for branding/features/languages.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateInstanceSettingsInput'
      responses:
        "200":
          description: Updated instance settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettings'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Неоторизиран достъп
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...